minor: additional tests for interaction around node lookup
A recent evaluation of the lookup code yielded an awareness that
while we're sort of following the overall community pattern of
testing what we expect in terms of patterns and behavior, really
needed just a little bit more in the way of testing.

While ultimately, these tests are really just exercising front
end validation, it is still important to check to have increased
assurance of a secure codebase.

Change-Id: Iaa917191e0f118f8828161174ea1fe8c55c8f4ee