Revert "ci: stable-only: explicitly pin centos build"
This reverts commit 86358c89e80cb2d078d1aef705cb1609e22da5bb.

Reason for revert: DIB appears to be still trying to merge filename in,
and as such we likely need to re-work our approach here.

Furthermore, explicitly force the redfish job to utilize just the wholedisk
test, instead of both the wholedisk and partition image build process, since
the partition image build test is easy to cause to fail.

Change-Id: I8c110deceda3a65e952cfb9b590d68b5d95efc16