Fix and document redfish metrics
Provides a complete documentation for metrics that the Redfish
management interface can collect.

The Power payload refers to InputRanges in a broken way: this field is a
list, but the code treats it as a singular resource. No hardware I have
access to provides it this way. Since input ranges are constants and
thus arguably don't qualify as runtime metrics, removing them instead of
fixing.

Change-Id: Ida1be1341346df917073e649a23a2f116b262e66