CI: Remove grub http boot from standalone CI test
Long story short, in some circles the EFI grub network boot
over http with VMs is regarded as unstable. What appears to
be happening, with service logs at least, is we get a HEAD
request (grub code always issues a HEAD request first to minimize
memory usage), and then re-requests the file contents.

So what we end up seeing on the grub side is:

error: Fail to receive a response! status=104

error: Fail to send a request! status=0x800000000000000f.

What appears to happen is things get gumed up in firmware and
the generally that means we can't run this test in CI.

Change-Id: I1471c9429b742abb250b9a3a910108f1711ad574