CI: stable/2023.1 - remove partition and older non-voting jobs
The partition image test job is known to be easy to cause to fail
because it is reliant upon CI scripting to try and make a partition
image during the job run, and it sometimes doesn't work which
increases the failure risk. Given this branch is moving un-maintained
in a few months, we can go ahead and remove this job with relatively
low risk at this point in time.

Also remove centos8, postgres, metalsmith, and non-voting bifrost jobs
to minimize resource waste.

Change-Id: I4cee2c24fc20227e84f7c25b5d24a4c9557b9614