Post-CVE follow-up fix - um/yoga and older
Somehow in the copying of various last minute changes I missed a
structural change to the logic where ultimately the originally
developed CVE fix patch was wrong, as the logic should have been

If we-re forcing raw:
- Then do so if we have deep image inspection disable and the
  qemu-info logic path tells us it is raw.
- Or when deep image inspection is enabled, and the detected format
  is *not* raw.

However, this ended with
- If we force raw and we have deep image inspection disabled and the
  image would be converted to raw
- Or deep image inspection is enabled and we detect the format of
  the image as *not* raw.

Anyhow, the difference is largely "force_raw" getting applied
across the board, because some operators *don't* force raw which
was the variation from testing  of the changes.

In any event, the original intended logic is correct, and should
resolve cases where the image is always getting converted for some
operators.

This patch will need to be backported across xena, wallaby,
and victoria.

Change-Id: I7baa03de2a26a05aa20bf88f0fdf051fc8be14fd