Refactoring: get rid of AgentDeployMixin
Its existence is probably a legacy of the iSCSI deploy times. Currently,
we have 4 different base classes/mixins in agent_base, which is
confusing even for a long-term contributor like me. AgentDeployMixin is
only used in CustomAgentDeploy, so it makes sense to get rid of it to
simplify the code navigation.

All deploy steps are moved to CustomAgentDeploy. Two two helper methods,
prepare_instance_to_boot and configure_local_boot are only used in
AgentDeploy, so moving them there.

Change-Id: Ib670571eb511d2f2e724ecfab1d2abb1ab471346