CI: Detail job removal permission
During the PTG, we discussed one of the challenges was keeping up on
removal of jobs *and* openly giving ourselves permission to remove
jobs which are *not* required after a certian point in time.

The critical aspect to this noting when/what we can clean up without
risk so we don't feel the need to keep things going forever when
there is not as much value to the overall health of the project.

Change-Id: I64f8f09c087d94376cbc32ef678a5da6595a805a