docs: begin making a general networking document
Apparently we had most of our networking context in the multitenancy
document, leading to confusion and also a lacking background context
which is needed for users to understand mechanics.

These changes have been broken into several different changes
because the file is going to be drastically different once we
done clarifying the documentation.

Change-Id: I7fd76a00d9bb344453fc1b214f170113d73fe9bc