Filter physnets when ports are pre-allocated to a segment
If an Ironic Node is connected to multiple physnets, and wants to
connect to a Neutron VIF whose network has segments in multiple
physnets it may be unclear which one to connect to.

If the VIF is already allocated to a subnet with a segment mapping,
this can be used to resolve the process to a physnet.

When a network uses Neutron routed segments, this patch first
checks the port's fixed_ips to see if a segment/physnet can be
determined from the subnet it is attached to. If this cannot be
identified (or fixed_ips is unset), the behaviour falls back to
returning all physnets available in the segmented network.

A similar change in Neutron was made in
I56b22820d29b2d57faf28a2f9b685ab0b2c924b4

Change-Id: Ie693257cdb8c44eeaf49cec9678de047f35d5221