Force legacy boot for metal3 integration job
Recently metal3-dev-env moved from bios to UEFI boot mode [1]
but the integration job on the metal3 CI still runs on
ubuntu jammy.
The location of the libvirt loader is different between
jammy and noble so we temporary use this workaround until
we support UEFI on ubuntu noble in metal3-dev-env.

[1] https://github.com/metal3-io/metal3-dev-env/pull/1326

Change-Id: I89caf341ac7078182e836f0fdd9bf08376837b20