API/Testing: Inspection rules migration
Migrates Inspector rules over to Ironic

This change defers the addition of the configuration options;
``supported_interfaces``, ``[auto_discovery]`` ``inspection_scope`` and
the ``default_scope`` as otherwise specified.

And, while the ``scope`` field may remain in the database for easier
migration, it's also outside of the 'scope' of this change as well,
and there's a chance we use the existing `traits` field in nodes as an
alternative for node-to-rule association instead.

A future follow-up should address these excluded implementations.

Change-Id: I6baf00273e63bb96e133f0cf5da6d8953f97af5a