move inspection hooks initialization
Moved the initialization of inspection hooks into the inspect_utils
module for future work of having multiple inspect interfaces utilize
this code. The hooks/base module defines the abstract interface for
hooks and provides code documentation. The rest of the directory has
other hooks so this shouldn't blend internal servers for executing the
hooks.

Change-Id: I756067000867f3eaa3a005f88e571a3666bea784
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>