oci: fix hang with 429 error code handling
While doing some additional testing, I've started to get 429 errors
from Quay which were causing my requests to hang. This was because
the built in retry logic with adapter use. As such, I removed the
adapter use and I now get a 429 error as expected and logged properly.

This was not caught with existing testing because it was getting
captured and held inside of urllib3 with the adapter usage.

Change-Id: I68a532a9765fbf90870ef4372b93738940eabd9e