Don't log PXE when your not doing PXE.
So... the neutron common code is modeled around PXE, and when you
go to do something with vmedia, you can get errors logged about not
having PXE enabled ports which is misleading.

The reality is the common code needs to base the decision on if all
ports need to be added or if just PXE enabled ports need to be added
based upon the loaded driver via configuration.

As such, now we consult the boot interface's capabilities field
and alternatively pull in the additional interfaces as appropriate
for virtual media users.

Which then also fixes the misleading error!

Closes-Bug: 2098791
Change-Id: I8c9d07e2ded75f138897ece6a67016a6f0020ce6