ci: restrict multinode job to multinode tests
We'll likely want to revert this one, but overall the multinode job
has been failing quite a bit because it is trying to run *all* of
the secenario jobs. While, that is not a big deal under normal
conditions, at end of development cycle, CI loads are much higher
and the failure rate can also be much higher as a result.

So for now, dial back the test to the most beneficial/needful
test which could be executed on the node.

Change-Id: Id4d050e1abd67f431026a2d813f97ffacf24407e