Enable atop on jobs
In order to stabilize CI we need to have clear understanding
on what is happening on the nodes. Right now we a blind cats
and trying to do predictions based on logs which is hard.
This patch enables atop installation on all jobs by default
which will provide a clear picture of what is happening
on the environment during tests, how many VMs are booted
concurrently etc.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/939578
Change-Id: I79c9c91e81c6a94f4e17191e05d78902dbb3b1f8