CI: Extend default timeouts slightly
Recently, logic was changed around network testing, but now we're seeing
test timeouts when the test is still running. In other words, we pushed
the time around and we started hitting the overall job timeouts.

As such, extend the window slightly on the default job so the test runs
can hopefully wrap up before the timeout trigger occurs.

Change-Id: Ie9d4fa6d747b93b2570f8abc13ef48c68e265644