[CI] Use bigger partition as work dir for metal3 job
In Rackspace env a bigger partition is available and configured
as /dev/xvde1 but not mounted.
We should use that as working dir for the metal3 job as the
standard one is too small and the job fails regularly when
running there.

This allows to make the metal3 job voting again.

Change-Id: I4d50b9c07367a5b1ad25887d87cf2e29ac6b4257