vncproxy: Use dedicated options for SSL
The [ssl] options from oslo.services are deprecated, because these have
been used to enable ssl in eventlet-based wsgi server, which is now
deprecated.

Introduce the dedicated options to [vnc] section to replace these
deprecated options.

Because the vnc console feature is not yet released, no release note
is added by this change.

Change-Id: If9d0e1413b46d7157bcd085280b80bf1e0e6355b