[CI] Fix libvirt logs collection in metal3 job
When the metal3 job fails the libvirt logs can't be collected
causing the entire log collection process to fail, de facto
making any troubleshooting process very difficult if not
impossible.
We need to ignore the error form the libvirt logs collection
task to allow the logs collection to finish correctly.

Change-Id: Ibbbb86457cbcf6c28392baad52478dc232d160bd