redfish: permit ipmitool-socat console for redfish usage
We've had a couple asks about enabling ipmitool-socat console usage
with the redfish hardware type.

This doesn't pickup parameters from driver_info to figure out the IPMI
username, password, and hostname/address, which turns out to be a rather
invasive change to ipmi handling code and test code as well.

Change-Id: I69dfeeec59317871c9f6a206966dd6a9d2ca9e3e