api: Ensure parameter transform happens early
The 'args.validate' decorator currently performs two functions:
transformation of serialised parameters into their unserialised formats
(i.e. a CSV string into a list of strings) and validation of the
parameters. While we are replacing use of the function for the latter,
we are retaining it for the former (at least for now).

The JSON Schema schemas rely on the transformed values, so we need to
ensure this transformation happens _before_ we run schema validation.
Ensure this is the case by looking for side-effects of the schema
validation decorators and erroring out if they're found (since that
implies they've already run). See [1] for more details.

[1] https://review.opendev.org/c/openstack/ironic/+/945943/2/ironic/api/controllers/v1/firmware.py#76

Change-Id: I642d582030eeacb0e060b3d375406e011abfb76a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>