Consider missing MTU invalid metadata
A bug was observed in nova behavior of interacting with ironic where
metadata could be missing in the payload to Ironic. While not great
it is also not awful. In any event, If there happens to be no DHCP
the lack of an MTU can be very problematic and result in a mismatch
between environment configuration and node state.

Closes-Bug: 2110322
Change-Id: Iea85ac4789d646dc85d0d8b22aa8e596b246234b