Fix redfish driver URL parsing
When ``redfish_address`` contains trailing slashes, the Redfish driver
would preserve them in the ``root_prefix``, causing sushy to error
attempting to instantiate the client using a potentially malformed URL
which results in a "Resource not found" error.

This may or not be a breaking change since we are now stripping trailing
slash(es) from custom paths, as well as explicitly setting a default
``root_prefix`` value if path is/becomes empty (inline comment suggested
that was the case but it was never implemented).

E.g: /test/redfish/v0/ -> /test/redfish/v0
     / -> /redfish/v1/
     /// -> /redfish/v1/

Closes-Bug: #2070791
Change-Id: I03f794350c6c0efd68131e6ede80c8a698983228