Change the novncproxy service code to handle eventlet-lessness
While exploring removing the invocation of eventlet.hubs.use_hub()
from the novncproxy service code, I discovered that we needed to
go ahead and optionally invoke the eventlet.hubs.use_hub()
depending on if the service launch has occured such that the runtime
has been monkey patched.

This is because the service does not operate properly when moneky
patched without the eventlet.hubs.use_hub() invocation. Similarly
if the eventlet monkey patch is removed entirely then the service
works as expected.

Change-Id: I976086d1645a3a8cc8b169adbeafdb2522452153
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>